// src/utils/types.ts
export interface KdbCertificate {
  id: number;
  environment: 'Prod' | 'Test' | 'Dev' | 'QA' | 'Lab' | 'Edu';
  server: string;
  kdbPath: string;
  alias: string;
  issuer: string;
  subject: string;
  validTo: string; // Expiry Date
  // Detailed fields
  trustStatus?: 'Trusted' | 'Expiring' | 'Expired' | 'Unknown';
  serialNumber?: string;
  version?: string;
  validFrom?: string;
  keyAlgorithm?: string;
  keySize?: string;
  signatureAlgorithm?: string;
  sha1?: string;
  sha256?: string;
  importDate?: string;
  cmsStashInfo?: string;
}